refactor: useOutsideClick 훅으로 빼기
