+++ src/hooks/useOutsideClick.tsx
@@ -1,0 +1,18 @@
+import { useEffect } from 'react';
+import type { RefObject } from 'react';
+
+export function useOutsideClick(ref: RefObject<HTMLElement>, callback: () => void) {
+  useEffect(() => {
+    const handleClickOutside = (event: MouseEvent) => {
+      if (ref.current && !ref.current.contains(event.target as Node)) {
+        callback();
+      }
+    };
+
+    document.addEventListener('click', handleClickOutside);
+
+    return () => {
+      document.removeEventListener('click', handleClickOutside);
+    };
+  }, []);
+}
